requirements.txt
requests>=2.28.2
Pillow
pycryptodome>=3.11.0
PySide6>=6.4.2
